+++ frontend/src/App.jsx
@@ -6,7 +6,6 @@ import RecipePage from "./components/recipePage/RecipePage";
 import { Routes, Route } from "react-router-dom";
 import PhotoAlbum from "./components/PhotoAlbum";
 import AddIngredient from "./components/ingredientPage/AddIngredient";
-import AllIngredients from "./components/ingredientPage/AllIngredients";
 import CreateShoppingLists from "./components/shoppingListPage/CreateShoppingLists";
 import ShoppingList from "./components/shoppingListPage/ShoppingList";
 import ShoppingListReview from "./components/shoppingListPage/ShoppingListReview";
@@ -30,7 +29,6 @@ function App() {
           <Route path="/shoppinglists" element={<ShoppingListReview userId={1} />} />
           <Route path="/photoalbum" element={<PhotoAlbum />} />
           <Route path="/add-ingredient" element={<AddIngredient userId={1}/>} />
-          <Route path="/all-ingredients" element={<AllIngredients userId={1}/>} />
           <Route path="/create-shoppinglists" element={<CreateShoppingLists userId={1}/>} />
         </Routes>
       </div>
